move getVCRSlowBlanking and vcr_sb_notifier from eAVSwitch to eAVControl
