+++ lib/python/StartEnigma.py
@@ -318,16 +318,16 @@ class PowerKey:
 		self.doAction(action="powertimerDeepStandby")
 
 
-class AutoScartControl:
+class AutoScartControl:  # TODO don't init of you have no Scart
 	def __init__(self, session):
 		self.force = False
-		self.current_vcr_sb = enigma.eAVSwitch.getInstance().getVCRSlowBlanking()
+		self.current_vcr_sb = enigma.eAVControl.getInstance().getVCRSlowBlanking()
 		if self.current_vcr_sb and config.av.vcrswitch.value:
 			self.scartDialog = session.instantiateDialog(Scart, True)
 		else:
 			self.scartDialog = session.instantiateDialog(Scart, False)
 		config.av.vcrswitch.addNotifier(self.recheckVCRSb)
-		enigma.eAVSwitch.getInstance().vcr_sb_notifier.get().append(self.VCRSbChanged)
+		enigma.eAVControl.getInstance().vcr_sb_notifier.get().append(self.VCRSbChanged)
 
 	def recheckVCRSb(self, configelement):
 		self.VCRSbChanged(self.current_vcr_sb)
@@ -350,7 +350,7 @@ def runScreenTest():
 			try:
 				with open(filename, "r") as fd:
 					line = fd.read().strip().replace("\0", "")
-					count = int(line) if isdecimal(line) else 0
+					count = int(line) if line.isdecimal() else 0
 				if count >= 3:
 					return False
 			except OSError as err:
